Separate callconv's for parfor and non-parfor lowering
